Update and rename main.cpp to GuessTheNumberGame.cpp
I have developed a kata and updated the code to support Random Test Cases.
